+++ cisdi-bid-ext/src/main/java/com/bid/ext/model/CcConstructWorkProg.java
@@ -7,7 +7,6 @@ import com.qygly.shared.BaseException;
 import com.qygly.shared.ad.entity.EntityTypeE;
 import com.qygly.shared.util.SharedUtil;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -112,7 +111,7 @@ public class CcConstructWorkProg {
          */
         public static final String CC_CONSTRUCT_WORK_ID = "CC_CONSTRUCT_WORK_ID";
         /**
-         * {"EN": "进展比例（%）", "ZH_CN": "实际进度比例", "ZH_TW": "进展比例（%）"}。
+         * {"EN": "进展比例（%）", "ZH_CN": "实际进度比例（%）", "ZH_TW": "进展比例（%）"}。
          */
         public static final String ACT_WBS_PCT = "ACT_WBS_PCT";
         /**
@@ -711,21 +710,21 @@ public class CcConstructWorkProg {
     }
 
     /**
-     * {"EN": "进展比例（%）", "ZH_CN": "实际进度比例", "ZH_TW": "进展比例（%）"}。
+     * {"EN": "进展比例（%）", "ZH_CN": "实际进度比例（%）", "ZH_TW": "进展比例（%）"}。
      */
-    private BigDecimal actWbsPct;
+    private Integer actWbsPct;
 
     /**
-     * 获取：{"EN": "进展比例（%）", "ZH_CN": "实际进度比例", "ZH_TW": "进展比例（%）"}。
+     * 获取：{"EN": "进展比例（%）", "ZH_CN": "实际进度比例（%）", "ZH_TW": "进展比例（%）"}。
      */
-    public BigDecimal getActWbsPct() {
+    public Integer getActWbsPct() {
         return this.actWbsPct;
     }
 
     /**
-     * 设置：{"EN": "进展比例（%）", "ZH_CN": "实际进度比例", "ZH_TW": "进展比例（%）"}。
+     * 设置：{"EN": "进展比例（%）", "ZH_CN": "实际进度比例（%）", "ZH_TW": "进展比例（%）"}。
      */
-    public CcConstructWorkProg setActWbsPct(BigDecimal actWbsPct) {
+    public CcConstructWorkProg setActWbsPct(Integer actWbsPct) {
         if (this.actWbsPct == null && actWbsPct == null) {
             // 均为null，不做处理。
         } else if (this.actWbsPct != null && actWbsPct != null) {
